Create MissingNumber
You were required to return the only missing number in an array of distinct integers where the integers were in range of 0 to length of array. For example :
nums = [3,0,1]
Output: 2
This is an array of size 3 and the elements expected are 0,1,2 and 3 but 2 is missing from the array so we return 2.